Issue #14937: Migrated section 4.8.8 Numeric Literals to chapter wise testing
